kernel: process: ret err if mod inactive process
If a process is in an inactive state (like it has faulted and the kernel
stopped it) then many of the operations in the `ProcessType` trait
should not be functional and instead should return errors. This makes
those functions return errors or explicitly say that they are not valid.

This is needed to help signal to capsules or other users of a process
that the process is not active and will not run again. For example, if a
capsule tries to call `.enter()` on a grant with an inactive process,
the `Kernel` object in sched.rs will find the matching process, but
trying to access one of its grant pointers will fail since it is not
valid to setup or access a grant on an inactive process. This error will
then get back to the capsule.